made the login process more intuitive, restricting how it's supposed to work
